Add missing tests and fix uint conversion
